Rename file walletsAndDeletegates to walletsAndValidators
