+++ src/modules/wallet/components/searchBarWallets/validators.js
@@ -2,7 +2,7 @@ import React from 'react';
 import Highlighter from 'react-highlight-words';
 import WalletVisual from '@wallet/components/walletVisual';
 import { truncateAddress } from '@wallet/utils/account';
-import styles from './walletsAndDeletegates.css';
+import styles from './walletsAndValidators.css';
 
 const Validators = ({
   validators,
