+++ src/modules/wallet/components/searchBarWallets/walletsAndValidators.css
